new central play button, settings nav refactor, remove change model option, remove old components, fix styles
