+++ src/assets/styles.sass
@@ -17,7 +17,7 @@ $widescreen-enabled: false
 $fullhd-enabled: false
 
 // Update some of Bulma's component variables
-$body-background-color: $beige-lighter
+$body-background-color: black
 $control-border-width: 2px
 $input-border-color: transparent
 $input-shadow: none
