+++ src/components/Camera.vue
@@ -10,19 +10,15 @@ div
           font-awesome-icon.stop(:icon="['fas', 'stop-circle']", @click="toggleRecord()")/
         #control-icon.double.no-recording(v-if="!isRecording")
             font-awesome-icon.inline.exit(:icon="['fas', 'times-circle']", @click="exitRecord()")/
-            font-awesome-icon.inline.resume(:icon="['fas', 'play-circle']", @click="toggleRecord()")/
       .settings-wrapper
-        #control-icon(v-if="!isSettings")
-          font-awesome-icon.settings(:icon="['fas', 'sliders-h']", @click="toggleSettings()")/
-        #control-icon(v-if="isSettings")
-          font-awesome-icon.settings(:icon="['fas', 'times']", @click="toggleSettings()")/
-    .menu(v-if="isSettings")
-      SettingNav/
+        SettingNav/
+    .play(v-if="!isRecording")
+      font-awesome-icon.resume(:icon="['fas', 'play-circle']", @click="toggleRecord()")/
 </template>
 
 <script>
-  import * as faceapi from 'face-api.js'
   import SettingNav from '@/components/SettingNav';
+  import * as faceapi from 'face-api.js'
   import {Action} from "@/utils/actions";
   const { globalShortcut } = require('electron').remote
   import {getConfiguration} from "@/utils/configuration";
@@ -48,6 +44,7 @@ div
         .then((stream) => {
           videoEl.srcObject = stream
           this.track = stream.getTracks()[0];
+          this.toggleRecord()
         })
         .catch((error) => {
           console.log('Error!', error); // eslint-disable-line no-console
@@ -92,8 +89,8 @@ div
             console.log('Error', error); // eslint-disable-line no-console
           })
 
-      setTimeout(() => this.onPlay())
-    },
+        setTimeout(() => this.onPlay())
+      },
       toggleRecord(){
         this.isRecording = !this.isRecording
         this.text = this.isRecording ? 'Stop' : 'Record'
@@ -132,4 +129,89 @@ div
     left: 0
     width: 270px
     height: 150px
+
+  .play
+    left: 33%
+    right: 33%
+    top: 40%
+    position: absolute
+    height: 33%
+    width: 33%
+    svg.resume
+      color: #28cd41
+      height: 100%
+      width: 100%
+      cursor: pointer
+    svg.resume:hover
+      color: rgb(0, 125, 27)
+
+  .wrapper
+    width: 100%
+    height: 100%
+    margin: 0
+    padding: 0
+    position: absolute
+    top: 0
+    left: 0
+    -webkit-app-region: drag
+
+  .top-bar
+    position: relative
+    background-color: rgba(0,0,0,0.5)
+    height: 45px
+    z-index: 2
+
+  .control-wrapper
+    position: relative
+    float: left
+    top: 10px
+    left: 10px
+    height: 25px
+    width: 50px
+
+  .settings-wrapper
+    position: relative
+    float: right
+    margin-top: 10px
+    margin-right: 10px
+    height: 25px
+    width: 100px
+
+  #control-icon:hover
+    cursor: pointer
+
+  #control-icon.double
+    width: 55px
+    svg
+      margin-right: 5px
+
+  #control-icon
+    -webkit-app-region: no-drag
+    color: rgb(255, 69, 58)
+    cursor: pointer
+    height: 25px
+    width: 25px
+    svg
+      border-radius: 50%
+      width: 100%
+      height: 100%
+    svg.inline
+      display: inline-block
+      width: 40%
+    svg.stop:hover
+      color: rgb(201, 52, 0)
+    svg.exit:hover
+      color: rgb(201, 52, 0)
+    svg.stop
+      color: rgb(255, 69, 58)
+    svg.exit
+      color: rgb(255, 69, 58)
+
+  .menu
+    position: absolute
+    right: 0
+    top: 0
+    z-index: 1
+    width: 200px
+    height: 100%
 </style>
