--- src/views/About.vue
@@ -1,5 +1,0 @@
-<template lang="pug">
-  .about
-    h1 This is an about page
-    router-link(:to="{ name: 'home' }") Back home
-</template>
